+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # @project-version@ (@build-time@)
 
-* Updated TOC version for WoW Patch 10.0.5.
+* Custom color for Name text now also supports transparency [Comment #6829].